# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Crossword App",
        "title": "Crossword API",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://crosswordbackend.onrender.com/",
    "basePath": "/",
    "paths": {
        "/admin/updateCrossword": {
            "post": {
                "description": "Allows admin to upload or replace the crossword JSON file for the current day.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Upload or update the crossword puzzle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin authentication key",
                        "name": "X-Admin-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Crossword JSON data",
                        "name": "crossword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crossword uploaded successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid admin key)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating crossword",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/updateScore": {
            "post": {
                "description": "Runs the daily crossword scoring process. Compares every user's stored crossword answers with the official solution and updates their total scores accordingly.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Compare user answers and update scores",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin authentication key",
                        "name": "X-Admin-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scores updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid admin key)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error loading solutions or updating scores",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/updateSolution": {
            "post": {
                "description": "Allows admin to upload the official solution JSON file after the puzzle day ends.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Upload or update the crossword solution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin authentication key",
                        "name": "X-Admin-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Solution JSON data",
                        "name": "solution",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solution uploaded successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid admin key)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating solution",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/google": {
            "post": {
                "description": "Verifies Google token, auto-registers user if new, returns JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate with Google",
                "parameters": [
                    {
                        "description": "Google ID Token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TokenSample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authenticated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid Google token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/crossword": {
            "get": {
                "description": "Returns crossword grid data and clues from JSON file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crossword"
                ],
                "summary": "Get today's crossword",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Crossword"
                        }
                    },
                    "500": {
                        "description": "Crossword not found or parse error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getsolution": {
            "post": {
                "description": "Fetches both the crossword question (grid, clues, etc.) and the corresponding solution for a given crossword ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crossword"
                ],
                "summary": "Get Crossword and Solution",
                "parameters": [
                    {
                        "description": "ID of the crossword to fetch",
                        "name": "crossword_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crossword and solution successfully fetched",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request or solution not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while reading files",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/leaderboard": {
            "get": {
                "description": "Returns a list of users sorted by score in descending order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Get leaderboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/submitcrossword": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stores or updates a user’s crossword answers. Requires JWT authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crossword"
                ],
                "summary": "Submit crossword answers",
                "parameters": [
                    {
                        "description": "Crossword answer data",
                        "name": "crossword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CrosswordAnswer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Answer stored",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Returns JWT token if user exists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Log in existing user",
                "parameters": [
                    {
                        "description": "User login info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "User not registered",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Creates a new user if not registered and returns a JWT token. If already registered, returns the user(no JWT token)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully registered",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Database or token error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Cell": {
            "type": "object",
            "properties": {
                "isBlank": {
                    "type": "boolean"
                },
                "numberAssociated": {
                    "type": "integer"
                }
            }
        },
        "model.Clues": {
            "type": "object",
            "properties": {
                "across": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UnitClue"
                    }
                },
                "down": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UnitClue"
                    }
                }
            }
        },
        "model.Crossword": {
            "type": "object",
            "properties": {
                "clues": {
                    "$ref": "#/definitions/model.Clues"
                },
                "columns": {
                    "type": "integer"
                },
                "crosswordID": {
                    "type": "integer"
                },
                "grid": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/model.Cell"
                        }
                    }
                },
                "rows": {
                    "type": "integer"
                }
            }
        },
        "model.CrosswordAnswer": {
            "description": "A user's crossword answer submission",
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UnitClue"
                    }
                },
                "crossword_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.TokenSample": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "your_token"
                }
            }
        },
        "model.UnitClue": {
            "description": "A singular clue with its index",
            "type": "object",
            "properties": {
                "clueID": {
                    "type": "integer",
                    "example": 1
                },
                "clueText": {
                    "type": "string",
                    "example": "APPLE"
                }
            }
        },
        "model.User": {
            "description": "User entity with unique email and optional crossword answer",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "score": {
                    "type": "integer",
                    "example": 0
                },
                "username": {
                    "type": "string",
                    "example": "JohnDoe"
                }
            }
        }
    }
}